chore: move CheckerReturn to poem_openapi::auth
